A lesser evil
I'd rather go for something super short like the Unlicense, but that one is possibly invalid because of poor definitions of the public domain. I don't get the point of the legalese, since it introduces meaningless subclauses: "including, but without limitation" is literally inconsequential. If I had the courage, I'd go for the "Very Simple Public License", a distilled MIT: 

Copyright (c) 2020 xdvom03
You can modify, distribute and use this software for any purpose without any
restrictions as long as you keep this copyright notice intact. The software is
provided without any warranty.

I don't care so much, because honestly, who would use this (and even if, who would stir up license trouble), but it is after all a school project, and I might get in trouble with opinionated people, and the actual validity of the license might end up less important than the fact that the license does not appear in the default choice window. So, CC0 it is.

I don't have any trademarks or patents attached, so here's the TLDR (and intended meaning):

Do whatever you want with what you find here. Don't be surprised by whatever ends up happening, for there's no warranty.