Overhaul data using crawl
Crawled useful sites again, added badly misclassed files. Moved folders around to make the number of choices generally smaller.
This creates a training data overhang; the next step is to analyse that data to find out similarities behind the underlying classes (as per #46),
and to improve the UX again (both the UI as per #60 and others, and the processing speeds as per #75 and #26 (in that order)).
First next agenda item, though, is most definitely #77.