Work towards executability
Resolves #85 (but without the tacked-on problems like display, as described in issue).

Removes features that are broken in newest LTK, thus resolving #90. The colour scheme in explainers is
not so necessary, but the frames that used to distinguish groups of widgets are now gone, which makes
 the GUI harder to read.

The crawler was improved to be faster and use its entire scorebase in deciding where to crawl.

The export system might be used instead of generated data: everything fits into megabytes anyway,
 so it might as well be in one file. We load it, then work in memory. This becomes #93.

Fixes #92 - such links are not followed.